resuscitate mcjs_inspector as mcjs_tools
This is supposed to be the new iteration of development tools for mcjs.

The current vision is a local web application that serves as development
aide for *mcjs* itself (not for developing JavaScript code!).

This commit introduces `mcjs_tools` crate, and the first iteration of
the main page design (currently viewable in a browser by running and
connecting to mockup-server.js).