mcjs_vm: add tool `scan_eval`
I'm using this to collect some data on how `eval` is used in the test262
test suite, to see whether to actually implement it or not.

The motivation is that there are some hacks I can do to avoid
implementing eval in certain situations. For example, if the only argument is a
string literal (e.g. `eval("some(code)")`), then I can just parse the
string and splice the resulting AST in place of the call.