mcjs_test262: use version 'dirty' when the working tree is dirty
I've decided it's not worth it to track _which_ commit I was building on
when the working tree is dirty.  The only case where this happens is
when I'm actively working on the code, and all I want is to keep
discarding the last test run and re-do it.

Before this change, every old '*-dirty' test run lingered around, and I
had to delete it manually.  Just no advantages to it.