interpreter: save IID at each step
This is sensibly more inefficient (requires pointer chasing at each
single instruction), but prevents a whole number of bugs where the IID
stored in InterpreterData was out-of-sync with the `iid` stack variable
in `run_regular`.

Previously, `iid` had to be manually stored in InterpreterData at
strategic times, such as every time we returned an Err. But there are
way more places where this turns out to be required than I initially
expected, so for now I'm storing it at the beginning of literally every
instruction and calling it a day.